+++ R/peru_transparencia_economica.R
@@ -1,0 +1,3 @@
+peru_transparencia_economica <- function(modulo = "gasto", actualizacion = "diaria") {
+    transpaeco(modulo = modulo, actualizacion = actualizacion)
+}