+++ R/set_donde_se_gasta.R
@@ -1,7 +1,0 @@
-#' @inheritParams elegir_donde_se_gasta
-set_donde_se_gasta <- function(departamento_meta = NULL) {
-    list(
-        departamento_meta = departamento_meta
-    ) |>
-        purrr::discard(is.null)
-}